requirements.txt
psycopg2
djangorestframework
django-rest-swagger
pyyaml-5.3.1